Add files via upload
This is a MATLAB code primarily developed for the contention window optimization of WLAN. A Genetic Fuzzy mechanism has been applied and the results generated by using this code have been published in Wireless Network Journal in 2021. The title of the paper is "A genetic fuzzy contention window optimization approach for IEEE 802.11 WLANs".  Due to the capacity issue, all the log files have been removed (6336 files). A user when compiling the source code, may easily generated the log files.